# src/main/java/org/taller01/catalogservice/web/ProductController.java
package org.taller01.catalogservice.web;

import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import java.util.List;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;
import org.taller01.catalogservice.domain.Product;
import org.taller01.catalogservice.dto.ProductCreateDto;
import org.taller01.catalogservice.dto.ProductUpdateDto;
import org.taller01.catalogservice.service.ProductService;

@Tag(name = "Catalog", description = "Products catalog management API")
@RestController
@RequestMapping("/api/products")
@RequiredArgsConstructor
public class ProductController {

    private final ProductService service;

    @GetMapping
    public List<Product> list() {
        return service.findAll();
    }

    @GetMapping("/{id}")
    public Product get(@PathVariable String id) {
        return service.findById(id);
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    public Product create(@RequestBody @Valid ProductCreateDto dto) {
        return service.create(dto);
    }

    @PutMapping("/{id}")
    public Product update(@PathVariable String id, @RequestBody @Valid ProductUpdateDto dto) {
        return service.update(id, dto);
    }

    @DeleteMapping("/{id}")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void delete(@PathVariable String id) {
        service.delete(id);
    }
}